tests/requirements.txt
black==20.8b1
coverage==5.3
flake8==3.8.4
flake8-quotes==3.2.0
isort==5.6.4
pycodestyle==2.6.0
pyflakes==2.2.0
pytest-asyncio==0.14.0
pytest-cov==2.10.1
pytest-mock==3.3.1
pytest-sugar==0.9.4
twine==3.2.0
